Made Quai CPU Miner a CLI app
To go with the new cli app, have added 2 flags, region and zone.

Main objective of this refactor was the ease of sanity checks and inheriting all the
cool features available in the cli library. Before we start integrating
new proxy and other features into the Miner wanted to make the Miner a
cli app.